+++ TestSystemAPI/src/main/java/com/app/service/testing/TypeOfQuestionService.java
@@ -2,6 +2,9 @@ package com.app.service.testing;
 
 import com.app.model.TypeOfQuestion;
 import com.app.repository.TypeOfQuestionRepository;
+import exceptions.QuestionTypeAlreadyExistsException;
+import exceptions.TestAlreadyExistsException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,24 +19,36 @@ public class TypeOfQuestionService {
         this.typeOfQuestionRepository = typeOfQuestionRepository;
     }
 
-    public void addTypeOfQuestion(String name) {
-        typeOfQuestionRepository.save(new TypeOfQuestion(name));
+    public TypeOfQuestion addTypeOfQuestion(String name) throws QuestionTypeAlreadyExistsException {
+        if(typeOfQuestionRepository.findByName(name).isPresent()) {
+            throw new QuestionTypeAlreadyExistsException("this question type is already exists");
+        }else {
+            return typeOfQuestionRepository.save(new TypeOfQuestion(name));
+        }
+
     }
 
-    //SET
-    public void setTypeOfQuestion(Long id, String name) {
-        TypeOfQuestion typeOfQuestion = typeOfQuestionRepository.getById(id);
-        typeOfQuestion.setName(name);
+    public TypeOfQuestion setTypeOfQuestion(Long id, String name) throws EntityNotFoundException {
+        if(typeOfQuestionRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("question type not found. Wrong id");
+        }else {
+            TypeOfQuestion typeOfQuestion = typeOfQuestionRepository.getById(id);
+            typeOfQuestion.setName(name);
+            return typeOfQuestionRepository.save(typeOfQuestion);
+        }
     }
 
-    //DELETE
-    public void deleteTypeOfQuestion(Long id) {
-        typeOfQuestionRepository.deleteById(id);
+    public void deleteTypeOfQuestion(Long id) throws EntityNotFoundException {
+        if(typeOfQuestionRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("question type not found. Wrong id");
+        }else {
+            typeOfQuestionRepository.deleteById(id);
+        }
+
     }
 
-    //SHOW
-    public List<String> showAllNamesOfType() {
+    public List<TypeOfQuestion> getAllNamesOfType() {
         List<TypeOfQuestion> typeObj = typeOfQuestionRepository.findAll();
-        return typeObj.stream().map(TypeOfQuestion::getName).collect(Collectors.toList());
+        return typeObj;
     }
 }
